Support bf16-mixed in random_rotation_matrix()

+++ zatom/utils/training_utils.py
@@ -499,7 +499,10 @@ def random_rotation_matrix(validate: bool = False, **tensor_kwargs: Any) -> torc
 
     if validate:
         assert torch.allclose(
-            rot_mat @ rot_mat.T, torch.eye(3, device=rot_mat.device), atol=1e-5, rtol=1e-5
+            rot_mat @ rot_mat.T,
+            torch.eye(3, device=rot_mat.device, dtype=rot_mat.dtype),
+            atol=1e-5,
+            rtol=1e-5,
         ), "Not a rotation matrix."
 
     return rot_mat
